Create platformer-realistic
I'm not sure if that's the best title.. But I wanted to save this somehow anyway after playing around with this.  Basically I made the movement more like "mario" and added some logic to the controls.  The character now doesnt stop right away as soon as you stop, but slows down (quickly), holding right/left and jumping makes for a longer jump, and tapping right/left and jumping makes for a short jump.  It's not perfect but it feels really good (to me at least) - now to add delta timing???