Fixed inconsistent behavior.
The arguments to Phaser.Tween.to are:
to(properties, duration, ease, autoStart, delay, repeat, yoyo) → {Phaser.Tween}

The original example's code had "false" as the last parameter, which must be a simple mistake as "repeat" is supposed to be a number instead of a boolean. A commenter had noted that the tween animation repeats forever with "repeat" set to false, whereas it should be set to "0".

Nothing fancy, just fixing an inconsistency.